pyproject.toml
[project]
name = "rishta-app-whatsapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.3",
    "dotenv>=0.9.9",
    "openai-agents>=0.2.3",
]
